Add controller methods
Add addTask()
Add removeTask()
Remove getTask as unnecessary one

+++ src/main/java/com/kalmykov/kanban/dao/TaskDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
-public interface TaskDao extends CrudRepository<Task, Integer> {
+public interface TaskDao extends CrudRepository<Task, Long> {
 }
